layout: put cards in card panel
