+++ src/assets/styles/map.scss
@@ -29,5 +29,5 @@
   z-index: 1;
   top: 0;
   bottom: 0;
-  overflow-x: hidden;
+  overflow: hidden;
 }
